requirements.txt
flask>=3.0.0
pyyaml>=6.0
sentence-transformers>=2.2.0
torch>=2.0.0
